# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Son Of Orakzai NGO API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Member {
  id: number;
  name: string;
  fatherName: string;
  cnic: string;
  phone: string;
  email: string;
  location: string;
  profession: string;
  skills: string;
  interest: string;
  /** @nullable */
  message?: string | null;
  status: string;
  createdAt: string;
}

export interface CreateMemberBody {
  name: string;
  fatherName: string;
  cnic: string;
  phone: string;
  email: string;
  location: string;
  profession: string;
  skills: string;
  interest: string;
  /** @nullable */
  message?: string | null;
}

export interface UpdateMemberBody {
  status?: string;
  name?: string;
  profession?: string;
}

export interface MembersStats {
  total: number;
  approved: number;
  pending: number;
  rejected: number;
}

export interface BlogPost {
  id: number;
  title: string;
  excerpt: string;
  content: string;
  category: string;
  author: string;
  /** @nullable */
  imageUrl?: string | null;
  publishedAt: string;
  createdAt: string;
}

export interface CreateContactBody {
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  subject: string;
  message: string;
}

export interface ContactMessage {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  subject: string;
  message: string;
  createdAt: string;
}

export interface ImpactStats {
  totalMembers: number;
  studentsTrainedCount: number;
  healthCasesCount: number;
  districtsReached: number;
  volunteerCount: number;
  projectsCompleted: number;
}

export type ListMembersParams = {
  status?: string;
  search?: string;
};

export type ListBlogPostsParams = {
  category?: string;
};