# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Son Of Orakzai NGO API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all members
 */
export const ListMembersQueryParams = zod.object({
  status: zod.coerce.string().optional(),
  search: zod.coerce.string().optional(),
});

export const ListMembersResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  fatherName: zod.string(),
  cnic: zod.string(),
  phone: zod.string(),
  email: zod.string(),
  location: zod.string(),
  profession: zod.string(),
  skills: zod.string(),
  interest: zod.string(),
  message: zod.string().nullish(),
  status: zod.string(),
  createdAt: zod.string(),
});
export const ListMembersResponse = zod.array(ListMembersResponseItem);

/**
 * @summary Submit membership application
 */
export const CreateMemberBody = zod.object({
  name: zod.string(),
  fatherName: zod.string(),
  cnic: zod.string(),
  phone: zod.string(),
  email: zod.string(),
  location: zod.string(),
  profession: zod.string(),
  skills: zod.string(),
  interest: zod.string(),
  message: zod.string().nullish(),
});

/**
 * @summary Get a single member
 */
export const GetMemberParams = zod.object({
  id: zod.coerce.number(),
});

export const GetMemberResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  fatherName: zod.string(),
  cnic: zod.string(),
  phone: zod.string(),
  email: zod.string(),
  location: zod.string(),
  profession: zod.string(),
  skills: zod.string(),
  interest: zod.string(),
  message: zod.string().nullish(),
  status: zod.string(),
  createdAt: zod.string(),
});

/**
 * @summary Update member status
 */
export const UpdateMemberParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateMemberBody = zod.object({
  status: zod.string().optional(),
  name: zod.string().optional(),
  profession: zod.string().optional(),
});

export const UpdateMemberResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  fatherName: zod.string(),
  cnic: zod.string(),
  phone: zod.string(),
  email: zod.string(),
  location: zod.string(),
  profession: zod.string(),
  skills: zod.string(),
  interest: zod.string(),
  message: zod.string().nullish(),
  status: zod.string(),
  createdAt: zod.string(),
});

/**
 * @summary Get member statistics overview
 */
export const GetMembersStatsResponse = zod.object({
  total: zod.number(),
  approved: zod.number(),
  pending: zod.number(),
  rejected: zod.number(),
});

/**
 * @summary List blog posts
 */
export const ListBlogPostsQueryParams = zod.object({
  category: zod.coerce.string().optional(),
});

export const ListBlogPostsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  excerpt: zod.string(),
  content: zod.string(),
  category: zod.string(),
  author: zod.string(),
  imageUrl: zod.string().nullish(),
  publishedAt: zod.string(),
  createdAt: zod.string(),
});
export const ListBlogPostsResponse = zod.array(ListBlogPostsResponseItem);

/**
 * @summary Get a blog post
 */
export const GetBlogPostParams = zod.object({
  id: zod.coerce.number(),
});

export const GetBlogPostResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  excerpt: zod.string(),
  content: zod.string(),
  category: zod.string(),
  author: zod.string(),
  imageUrl: zod.string().nullish(),
  publishedAt: zod.string(),
  createdAt: zod.string(),
});

/**
 * @summary Submit contact form
 */
export const SubmitContactBody = zod.object({
  name: zod.string(),
  email: zod.string(),
  phone: zod.string().nullish(),
  subject: zod.string(),
  message: zod.string(),
});

/**
 * @summary Get impact statistics
 */
export const GetImpactStatsResponse = zod.object({
  totalMembers: zod.number(),
  studentsTrainedCount: zod.number(),
  healthCasesCount: zod.number(),
  districtsReached: zod.number(),
  volunteerCount: zod.number(),
  projectsCompleted: zod.number(),
});